requirements.txt
asgiref==3.7.2
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
coverage==7.4.3
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg==3.1.18
psycopg2==2.9.9
pycparser==2.22
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
